Moved MLTaskManager out of index.js to its own file
